mva-sidecar/requirements.txt
numpy>=1.26.0
scipy>=1.11.0
scikit-learn>=1.4.0
requests>=2.31.0
prometheus-client>=0.19.0
PyYAML>=6.0.1
urllib3>=2.0.0
